% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ComputeAgeHeapingScores.R
\name{ComputeAgeHeapingScores}
\alias{ComputeAgeHeapingScores}
\title{Compute summaries of age heaping (wrapper for DemoTools::check_heaping_*)}
\usage{
ComputeAgeHeapingScores(
  data,
  name.disaggregations,
  name.age,
  name.sex,
  name.males,
  name.females,
  name.date1,
  name.date2,
  name.population.year1,
  name.population.year2,
  roughness.age.min = NULL,
  roughness.age.max = NULL,
  sawtooth.age.min = NULL,
  sawtooth.age.max = NULL,
  Whipple.age.min = NULL,
  Whipple.age.max = NULL,
  Whipple.digit = NULL,
  Myers.age.min = NULL,
  Myers.age.max = NULL,
  Noumbissi.age.min = NULL,
  Noumbissi.age.max = NULL,
  Noumbissi.digit = NULL
)
}
\arguments{
\item{data}{asdf}

\item{name.disaggregations}{asdf}

\item{name.age}{asdf}

\item{name.sex}{asdf}

\item{name.males}{asdf}

\item{name.females}{asdf}

\item{name.date1}{asdf}

\item{name.date2}{asdf}

\item{name.population.year1}{asdf}

\item{name.population.year2}{asdf}

\item{roughness.age.min=NULL}{asdf}

\item{roughness.age.max=NULL}{asdf}

\item{sawtooth.age.min=NULL}{asdf}

\item{sawtooth.age.max=NULL}{asdf}

\item{Whipple.age.min=NULL}{asdf}

\item{Whipple.age.max=NULL}{asdf}

\item{Whipple.digit=NULL}{asdf}

\item{Myers.age.min=NULL}{asdf}

\item{Myers.age.max=NULL}{asdf}

\item{Noumbissi.age.min=NULL}{asdf}

\item{Noumbissi.age.max=NULL}{asdf}

\item{Noumbissi.digit=NULL}{asdf}
}
\description{
This function serves as a wrapper for five DemoTools functions -- 
check_heaping_roughness(), check_heaping_sawtooth(), check_heaping_whipple(), check_heaping_myers(), check_heaping_noumbissi() -- 
and returns the five summary statistics within specified levels of disaggregation in a given dataset 
(which must contain single-year age counts) separately for males and females
}
