% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ecuador_five_year_ages_national}
\alias{ecuador_five_year_ages_national}
\title{Five-year age groups example dataset for package \code{SubNationalCRVS}}
\format{
A data frame with X rows and Y variables
\itemize{
  \item asdf
}
}
\usage{
ecuador_five_year_ages_national
}
\description{
Add description / citations to Andres Peralta and Ecuador govt
}
\keyword{datasets}
