% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ecuador_single_year_ages_national}
\alias{ecuador_single_year_ages_national}
\title{Single-year-age example dataset for package \code{SubnationalCRVS}}
\format{A data frame with X rows and Y variables
\itemize{
  \item asdf
}}
\usage{
ecuador_single_year_ages_national
}
\description{
Add description / citations to Andres Peralta and Ecuador govt
}
\keyword{datasets}
