% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EstimateDDM.R
\name{EstimateDDM}
\alias{EstimateDDM}
\title{Estimate death registration between two Census years with a wrapper of the ddm() function from the DDM}
\usage{
EstimateDDM(
  data,
  name.disaggregations,
  name.age,
  name.sex,
  name.males,
  name.females,
  name.date1,
  name.date2,
  name.population.year1,
  name.population.year2,
  name.deaths,
  deaths.summed,
  show_age_range_sensitivity = TRUE,
  min.age.in.search = 15,
  max.age.in.search = 75,
  min.number.of.ages = 8,
  exact.ages.to.use = NULL,
  largest.lower.limit.sensitivity = 45,
  smallest.upper.limit.sensitivity = 50,
  life.expectancy.in.open.group = NULL
)
}
\arguments{
\item{data}{asdf}

\item{name.disaggregations}{asdf}

\item{name.age}{asdf}

\item{name.sex}{asdf}

\item{name.males}{asdf}

\item{name.females}{asdf}

\item{name.date1}{asdf}

\item{name.date2}{asdf}

\item{name.population.year1}{asdf}

\item{name.population.year2}{asdf}

\item{deaths.summed}{asdf}

\item{show_age_range_sensitivity}{asdf}

\item{min.age.in.search}{asdf}

\item{max.age.in.search}{asdf}

\item{min.number.of.ages}{asdf}

\item{exact.ages.to.use}{asdf}

\item{largest.lower.limit.sensitivity}{asdf}

\item{smallest.upper.limit.sensitivity}{asdf}

\item{life.expectancy.in.open.group}{asdf}
}
\description{
asdf
}
