% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\name{example}
\alias{example}
\title{Example}
\usage{
example(x)
}
\arguments{
\item{x}{An example parameter}
}
\value{
A example result
}
\description{
deqn ASCII example
}
\details{
\deqn{ \sigma = \sqrt{ \frac{Z}{n} \sum
 \left[ \textstyle\frac{1}{2}\displaystyle
   \left( \log \frac{H_i}{L_i} \right)^2  - (2\log 2-1)
   \left( \log \frac{C_i}{O_i} \right)^2 \right] }
}{sqrt(N/n * runSum(0.5 * log(OHLC[,2]/OHLC[,3])^2 -
          (2*log(2)-1) * log(OHLC[,4]/OHLC[,1])^2, n))}
}
\author{
Joshua Ulrich
}
\keyword{ts}
