% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotAgeHeapingScores.R
\name{PlotAgeHeapingScores}
\alias{PlotAgeHeapingScores}
\title{Plot age heaping scores}
\usage{
PlotAgeHeapingScores(
  data,
  name.disaggregations,
  name.age,
  name.sex,
  name.males,
  name.females,
  name.date1,
  name.date2,
  name.population.year1,
  name.population.year2,
  name.national = NULL,
  show.size.population = TRUE,
  roughness.age.min = NULL,
  roughness.age.max = NULL,
  Whipple.age.min = NULL,
  Whipple.age.max = NULL,
  Whipple.digit = NULL,
  Myers.age.min = NULL,
  Myers.age.max = NULL,
  label.subnational.level = name.disaggregations,
  base.size = 12,
  fig.nrow = 1,
  fig.ncol = 1,
  save.name.plots = NULL,
  plots.dir = ""
)
}
\arguments{
\item{data}{data frame that contains at least seven columns representing: (1) single-year age,
(2) sex,
(3, 4) population counts collected at two different time points (typically adjacent Census years)
(5, 6) dates of two different time points
(7) the level of subnational disaggregation in addition to sex (e.g. a geographic unit such as a province/state, 
a sociodemographic category such as education level, or combinations thereof).}

\item{name.disaggregations}{Character string providing the name of the variable in `data` that represents the levels of subnational disaggregation}

\item{name.age}{Character string providing the name of the variable in `data` that represents age}

\item{name.sex}{Character string providing the name of the variable in `data` that represents sex}

\item{name.males}{Character string providing the name of the value of `name.sex` variable that represents males}

\item{name.females}{Character string providing the name of the value of `name.sex` variable that represents females}

\item{name.date1}{Character string providing the name of the variable in `data` that represents the earlier time period}

\item{name.date2}{Character string providing the name of the variable in `data` that represents the later time period}

\item{name.population.year1}{Character string providing the name of the variable in `data` that represents the population count in the earlier time period}

\item{name.population.year2}{Character string providing the name of the variable in `data` that represents the population count in the later time period}

\item{name.national}{A character string providing the value of `name.disaggregations` variable that indicates national-level results (e.g. "Overall" or "National"). Defaults to NULL, implying `name.disaggregations` variable in `data` only includes values for subnational levels. Defaults to NULL}

\item{show.size.population}{A logical indixating whether the size of plotted points should vary according to the total population size in the second data year. Defaults to TRUE}

\item{label.subnational.level}{A character label for the axis showing the level of subnational disaggregation present in the data. Defaults to `name.disaggregations`}

\item{base.size}{A numeric fed to `ggplot2::theme_classic(base_size)` for the plot of point estimates. Defaults to 12}

\item{fig.nrow}{An integer fed to `gridExtra::arrangeGrob(nrow)` to indicate how many rows on each page should be used to display the 5 plots}

\item{fig.ncol}{An integer fed to `gridExtra::arrangeGrob(ncol)` to indicate how many columns on each page should be used to display the 5 plots}

\item{save.name.plots}{A character specifying a custom file name for the plots saved on the local file system. Defaults to NULL, which combines `name.disaggregations` and the current date}

\item{plots.dir}{A character specifying the directory where plots should be saved. Defaults to "", saving the plots in the working directory}

\item{roughness.age.min=NULL}{Equivalent to the `ageMin` argument of `Demotools::check_heaping_roughness`. Defaults to NULL, which then uses the `DemoTools` default of 20}

\item{roughness.age.max=NULL}{Equivalent to the `ageMax` argument of `Demotools::check_heaping_roughness`. Defaults to NULL, which then uses the `DemoTools` default of the highest age that is a multiple of 10}

\item{Whipple.age.min=NULL}{Equivalent to the `ageMin` argument of `Demotools::check_heaping_whipple`. Defaults to NULL, which then uses the `DemoTools` default of 25}

\item{Whipple.age.max=NULL}{Equivalent to the `ageMax` argument of `Demotools::check_heaping_whipple`. Defaults to NULL, which then uses the `DemoTools` default of 65}

\item{Whipple.digit=NULL}{Equivalent to the `digit` argument of `Demotools::check_heaping_whipple`. Defaults to NULL, which then uses the `DemoTools` default of c(0, 5)}

\item{Myers.age.min=NULL}{Equivalent to the `ageMin` argument of `Demotools::check_heaping_myers`. Defaults to NULL, which then uses the `DemoTools` default of 10}

\item{Myers.age.max=NULL}{Equivalent to the `ageMax` argument of `Demotools::check_heaping_myers`. Defaults to NULL, which then uses the `DemoTools` default of 89}
}
\description{
asdf
}
\examples{
age_heaping_plotting <- PlotAgeHeapingScores(data=ecuador_single_year_ages
                                                  name.disaggregations="province_name",
                                                  name.males="m",
                                                  name.females="f",
                                                  name.age="age",
                                                  name.sex="sex",
                                                  name.date1="date1",
                                                  name.date2="date2",
                                                  name.population.year1="pop1",
                                                  name.population.year2="pop2",
                                                  plots.dir="Plots/")
head(age_heaping_plotting)
}
