% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotDDM.R
\name{PlotDDM}
\alias{PlotDDM}
\title{Plot DDM estimates and their sensitivity to choice of age range}
\usage{
PlotDDM(
  ddm_results,
  size.text.sensitivity = 8,
  fig.nrow = 2,
  fig.ncol = 1,
  print.plot.point.estimates = TRUE,
  save.plot.point.estimates = TRUE,
  save.name.plot.point.estimates = NULL,
  print.plots.sensitivity = FALSE,
  save.plots.sensitivity = TRUE,
  save.name.plots.sensitivity = NULL,
  plots.dir = ""
)
}
\arguments{
\item{ddm_results}{asdf}
}
\description{
asdf
}
