% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotPotentialAgeHeaping.R
\name{PlotPotentialAgeHeaping}
\alias{PlotPotentialAgeHeaping}
\title{Plot counts of single-year ages to reveal potential age heaping}
\usage{
PlotPotentialAgeHeaping(
  data,
  name.disaggregations,
  name.age,
  name.sex,
  name.males,
  name.females,
  name.date1,
  name.date2,
  name.population.year1,
  name.population.year2,
  mark_multiples_of_5_disaggregated = FALSE,
  mark_multiples_of_5_overall = TRUE,
  ylim.disaggregated = NULL,
  ylim.overall = NULL,
  line.size.disaggregated = 0.6,
  line.size.overall = 0.6,
  fig.nrow.disaggregated = 3,
  fig.ncol.disaggregated = 2,
  fig.nrow.overall = 2,
  fig.ncol.overall = 1,
  print.disaggregated = TRUE,
  save.disaggregated = TRUE,
  save.name_disaggregated = NULL,
  print.overall = TRUE,
  save.overall = TRUE,
  save.name_overall = NULL,
  plots.dir = ""
)
}
\arguments{
\item{data}{sadf}

\item{name.disaggregations}{asdf}

\item{name.age}{asdf}

\item{name.sex}{asdf}

\item{name.males}{asdf}

\item{name.females}{asdf}

\item{name.date1}{asdf}

\item{name.date2}{asdf}

\item{name.population.year1}{asdf}

\item{name.population.year2}{asdf}

\item{mark_multiples_of_5_disaggregated}{asdf}

\item{mark_multiples_of_5_overall}{asdf}

\item{ylim.disaggregated}{asdf}

\item{ylim.overall}{asdf}

\item{line.size.disaggregated}{asdf}

\item{line.size.overall}{asdf}

\item{fig.nrow.disaggregated}{asdf}

\item{fig.ncol.disaggregated}{asdf}

\item{fig.nrow.overall}{asdf}

\item{fig.ncol.overall}{asdf}

\item{print.disaggregated}{asdf}

\item{save.disaggregated}{asdf}

\item{save.name_disaggregated}{asdf}

\item{print.overall}{asdf}

\item{save.overall}{asdf}

\item{save.name_overall}{asdf}

\item{plots.dir}{asdf}
}
\description{
asdf
}
