% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotSexRatios.R
\name{PlotSexRatios}
\alias{PlotSexRatios}
\title{Plot sex ratios}
\usage{
PlotSexRatios(
  data,
  name.disaggregations,
  name.age,
  name.sex,
  name.males,
  name.females,
  name.date1,
  name.date2,
  name.population.year1,
  name.population.year2,
  ylim.disaggregated = NULL,
  ylim.overall = NULL,
  line.size.disaggregated = 0.8,
  line.size.overall = 0.8,
  line.size = 1.1,
  fig.nrow = 3,
  fig.ncol = 2,
  print.disaggregated = TRUE,
  save.disaggregated = TRUE,
  save.name_disaggregated = NULL,
  print.overall = TRUE,
  save.overall = TRUE,
  save.name_overall = NULL,
  plots.dir = ""
)
}
\arguments{
\item{data}{sadf}

\item{name.disaggregations}{asdf}

\item{name.age}{asdf}

\item{name.sex}{asdf}

\item{name.males}{asdf}

\item{name.females}{asdf}

\item{name.date1}{asdf}

\item{name.date2}{asdf}

\item{name.population.year1}{asdf}

\item{name.population.year2}{asdf}
}
\description{
asdf
}
