% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ComputeAgeHeapingScores.R
\name{ComputeAgeHeapingScores}
\alias{ComputeAgeHeapingScores}
\title{Compute summaries of age heaping (wrapper for DemoTools::check_heaping_*)}
\usage{
ComputeAgeHeapingScores(
  data,
  name.disaggregations,
  name.age,
  name.sex,
  name.males,
  name.females,
  name.date1,
  name.date2,
  name.population.year1,
  name.population.year2,
  roughness.age.min = NULL,
  roughness.age.max = NULL,
  sawtooth.age.min = NULL,
  sawtooth.age.max = NULL,
  Whipple.age.min = NULL,
  Whipple.age.max = NULL,
  Whipple.digit = NULL,
  Myers.age.min = NULL,
  Myers.age.max = NULL,
  Noumbissi.age.min = NULL,
  Noumbissi.age.max = NULL,
  Noumbissi.digit = NULL
)
}
\arguments{
\item{data}{data frame that contains at least seven columns representing: (1) single-year age,
(2) sex,
(3, 4) population counts collected at two different time points (typically adjacent Census years)
(5, 6) dates of two different time points
(7) the level of subnational disaggregation in additino to sex (e.g. a geographic unit such as a province/state, 
a sociodemographic category such as education level, or combinations thereof).}

\item{name.disaggregations}{Character string providing the name of the variable in `data` that represents the levels of subnational disaggregation}

\item{name.age}{Character string providing the name of the variable in `data` that represents age}

\item{name.sex}{Character string providing the name of the variable in `data` that represents sex}

\item{name.males}{Character string providing the name of the value of `name.sex` variable that represents males}

\item{name.females}{Character string providing the name of the value of `name.sex` variable that represents females}

\item{name.date1}{Character string providing the name of the variable in `data` that represents the earlier time period}

\item{name.date2}{Character string providing the name of the variable in `data` that represents the later time period}

\item{name.population.year1}{Character string providing the name of the variable in `data` that represents the population count in the earlier time period}

\item{name.population.year2}{Character string providing the name of the variable in `data` that represents the population count in the later time period}

\item{roughness.age.min=NULL}{Equivalent to the `ageMin` argument of `Demotools::check_heaping_roughness`. Defaults to NULL, which then uses the `DemoTools` default of 20}

\item{roughness.age.max=NULL}{Equivalent to the `ageMax` argument of `Demotools::check_heaping_roughness`. Defaults to NULL, which then uses the `DemoTools` default of the highest age that is a multiple of 10}

\item{sawtooth.age.min=NULL}{Equivalent to the `ageMin` argument of `Demotools::check_heaping_sawtooth`. Defaults to NULL, which then uses the `DemoTools` default of 40}

\item{sawtooth.age.max=NULL}{Equivalent to the `ageMax` argument of `Demotools::check_heaping_sawtooth`. Defaults to NULL, which then uses the `DemoTools` default of the highest age that is a multiple of 10}

\item{Whipple.age.min=NULL}{Equivalent to the `ageMin` argument of `Demotools::check_heaping_whipple`. Defaults to NULL, which then uses the `DemoTools` default of 25}

\item{Whipple.age.max=NULL}{Equivalent to the `ageMax` argument of `Demotools::check_heaping_whipple`. Defaults to NULL, which then uses the `DemoTools` default of 65}

\item{Whipple.digit=NULL}{Equivalent to the `digit` argument of `Demotools::check_heaping_whipple`. Defaults to NULL, which then uses the `DemoTools` default of c(0, 5)}

\item{Myers.age.min=NULL}{Equivalent to the `ageMin` argument of `Demotools::check_heaping_myers`. Defaults to NULL, which then uses the `DemoTools` default of 10}

\item{Myers.age.max=NULL}{Equivalent to the `ageMax` argument of `Demotools::check_heaping_myers`. Defaults to NULL, which then uses the `DemoTools` default of 89}

\item{Noumbissi.age.min=NULL}{Equivalent to the `ageMin` argument of `Demotools::check_heaping_noumbissi`. Defaults to NULL, which then uses the `DemoTools` default of 20}

\item{Noumbissi.age.max=NULL}{Equivalent to the `ageMax` argument of `Demotools::check_heaping_noumbissi`. Defaults to NULL, which then uses the `DemoTools` default of 64}

\item{Noumbissi.digit=NULL}{Equivalent to the `digit` argument of `Demotools::check_heaping_noumbissi`. Defaults to NULL, which then uses the `DemoTools` default of 0}
}
\description{
This function serves as a wrapper for five DemoTools functions -- 
check_heaping_roughness(), check_heaping_sawtooth(), check_heaping_whipple(), check_heaping_myers(), check_heaping_noumbissi() -- 
and returns the five summary statistics within specified levels of disaggregation in a given dataset 
(which must contain single-year age counts) separately for males and females.
}
\examples{
ecuador_age_heaping_scores <- ComputeAgeHeapingScores(data=ecuador_age_tabulation,
                                              name.disaggregations="province",
                                              name.males="m",
                                              name.females="f",
                                              name.age="age",
                                              name.sex="sex",
                                              name.date1="date1",
                                              name.date2="date2",
                                              name.population.year1="pop1",
                                              name.population.year2="pop2")
head(ecuador_age_heaping_scores)
tail(ecuador_age_heaping_scores)
}
