% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ComputeSexRatios.R
\name{ComputeSexRatios}
\alias{ComputeSexRatios}
\title{Compute sex ratios for adjacent five-year age groups}
\usage{
ComputeSexRatios(
  data,
  name.disaggregations,
  name.age,
  name.sex,
  name.males,
  name.females,
  name.date1,
  name.date2,
  name.population.year1,
  name.population.year2
)
}
\arguments{
\item{data}{asdf}

\item{name.disaggregations}{asdf}

\item{name.age}{asdf}

\item{name.sex}{asdf}

\item{name.males}{asdf}

\item{name.females}{asdf}

\item{name.date1}{asdf}

\item{name.date2}{asdf}

\item{name.population.year1}{asdf}

\item{name.population.year2}{asdf}
}
\description{
asdf
}
