% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotAgeHeapingScores.R
\name{PlotAgeHeapingScores}
\alias{PlotAgeHeapingScores}
\title{Plot age heaping scores}
\usage{
PlotAgeHeapingScores(
  data,
  name.disaggregations,
  name.age,
  name.sex,
  name.males,
  name.females,
  name.date1,
  name.date2,
  name.population.year1,
  name.population.year2,
  roughness.age.min = NULL,
  roughness.age.max = NULL,
  sawtooth.age.min = NULL,
  sawtooth.age.max = NULL,
  Whipple.age.min = NULL,
  Whipple.age.max = NULL,
  Whipple.digit = NULL,
  Myers.age.min = NULL,
  Myers.age.max = NULL,
  fig.nrow = 3,
  fig.ncol = 1,
  print.plots = TRUE,
  save.plots = TRUE,
  save.name.plots = NULL,
  plots.dir = ""
)
}
\arguments{
\item{data}{sadf}

\item{name.disaggregations}{asdf}

\item{name.age}{asdf}

\item{name.sex}{asdf}

\item{name.males}{asdf}

\item{name.females}{asdf}

\item{name.date1}{asdf}

\item{name.date2}{asdf}

\item{name.population.year1}{asdf}

\item{name.population.year2}{asdf}

\item{roughness.age.min}{asdf}

\item{roughness.age.max}{asdf}

\item{sawtooth.age.min}{asdf}

\item{sawtooth.age.max}{asdf}

\item{Whipple.age.min}{asdf}

\item{Whipple.age.max}{asdf}

\item{Whipple.digit}{asdf}

\item{Myers.age.min}{asdf}

\item{Myers.age.max}{asdf}

\item{fig.nrow}{asdf}

\item{fig.ncol}{asdf}

\item{print.plots}{asdf}

\item{save.plots}{asdf}

\item{save.name.plots}{asdf}

\item{plots.dir}{asdf}
}
\description{
asdf
}
