% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotPotentialAgeHeaping.R
\name{PlotPotentialAgeHeaping}
\alias{PlotPotentialAgeHeaping}
\title{Plot counts of single-year ages to reveal potential age heaping}
\usage{
PlotPotentialAgeHeaping(
  data,
  name.disaggregations,
  name.age,
  name.sex,
  name.males,
  name.females,
  name.date1,
  name.date2,
  name.population.year1,
  name.population.year2,
  mark_multiples_of_5_disaggregated = FALSE,
  mark_multiples_of_5_overall = TRUE,
  ylim.disaggregated = NULL,
  ylim.overall = NULL,
  line.size.disaggregated = 0.6,
  line.size.overall = 0.6,
  fig.nrow.disaggregated = 3,
  fig.ncol.disaggregated = 2,
  fig.nrow.overall = 2,
  fig.ncol.overall = 1,
  print.disaggregated = FALSE,
  save.disaggregated = TRUE,
  save.name_disaggregated = NULL,
  print.overall = TRUE,
  save.overall = TRUE,
  save.name_overall = NULL,
  plots.dir = ""
)
}
\arguments{
\item{data}{data frame that contains at least seven columns representing: (1) single-year age,
(2) sex,
(3, 4) population counts collected at two different time points (typically adjacent Census years)
(5, 6) dates of two different time points
(7) the level of subnational disaggregation in additino to sex (e.g. a geographic unit such as a province/state, 
a sociodemographic category such as education level, or combinations thereof).}

\item{name.disaggregations}{Character string providing the name of the variable in `data` that represents the levels of subnational disaggregation}

\item{name.age}{Character string providing the name of the variable in `data` that represents age}

\item{name.sex}{Character string providing the name of the variable in `data` that represents sex}

\item{name.males}{Character string providing the name of the value of `name.sex` variable that represents males}

\item{name.females}{Character string providing the name of the value of `name.sex` variable that represents females}

\item{name.date1}{Character string providing the name of the variable in `data` that represents the earlier time period}

\item{name.date2}{Character string providing the name of the variable in `data` that represents the later time period}

\item{name.population.year1}{Character string providing the name of the variable in `data` that represents the population count in the earlier time period}

\item{name.population.year2}{Character string providing the name of the variable in `data` that represents the population count in the later time period}

\item{mark_multiples_of_5_disaggregated}{Logical indicating whether dashed vertical lines should be overlaid on the disaggregated (i.e. separate plot within each level of disaggregation) plots at ages that are multiples of 5. Defaults to FALSE}

\item{mark_multiples_of_5_overall}{Logical indicating whether dashed vertical lines should be overlaid on the overall (i.e. all levels of disaggregation are plotted on the same graph) plots at ages that are multiples of 5. Defaults to TRUE}

\item{ylim.disaggregated}{A vector with two numeric entries indicating the minimum and maximum values of the y-axis for the age-specific population counts plotted on a separate graph within each level of disaggregation. Default to NULL, which uses the smallest and largest counts within each level}

\item{ylim.overall}{A vector with two numeric entries indicating the minimum and maximum values of the y-axis for the overall age-specific population counts where all levels of disaggregation are plotted on the same graph. Defaults to NULL, which uses the smallest and largest counts in the entire dataset}

\item{line.size.disaggregated}{Numeric fed into ggplot2::geom_line(size)) for the disaggregated plots. Defaults to 0.6}

\item{line.size.overall}{Numeric fed into ggplot2::geom_line(size)) for the overall plot. Defaults to 0.6}

\item{fig.nrow.disaggregated}{An integer fed to `gridExtra::arrangeGrob(nrow)` to indicate how many rows should be used on each page to display the disaggregated plots. Defaults to 3}

\item{fig.ncol.disaggregated}{An integer fed to `gridExtra::arrangeGrob(ncol)` to indicate how many columns should be used on each page to display the disaggregated plots. Defaults to 2}

\item{fig.nrow.overall}{An integer fed to `gridExtra::arrangeGrob(nrow)` to indicate how many rows should be used on each page to display the overall plot. Defaults to 2}

\item{fig.ncol.overall}{An integer fed to `gridExtra::arrangeGrob(ncol)` to indicate how many columns should be used on each page to display the overall plot. Defaults to 1}

\item{print.disaggregated}{A logical indicating whether the disaggregated plots should be printed in the R session. Defaults to FALSE}

\item{save.disaggregated}{A logical indicating whether the disaggregated plots should be saved on the local file system. Defaults to TRUE}

\item{print.overall}{A logical indicating whether the overall plots should be printed in the R session. Defaults to TRUE}

\item{save.overall}{A logical indicating whether the overall plots should be saved on the local file system. Defaults to TRUE}

\item{plots.dir}{A character specifying the directory where plots should be saved. Defaults to "", saving the plots in the working directory}

\item{save.name.disaggregated}{A character specifying a custom file name for the disaggregated plots saved on the local file system. Defaults to NULL, which combines `name.disaggregations` and the current date}

\item{save.name.overall}{A character specifying a custom file name for the overall plots saved on the local file system. Defaults to NULL, which combines `name.disaggregations` and the current date}
}
\description{
asdf
}
\examples{
PlotPotentialAgeHeaping(data=ecuador_age_tabulation,
                        name.disaggregations="province_name_short",
                        name.males="m",
                        name.females="f",
                        name.age="age",
                        name.sex="sex",
                        name.date1="date1",
                        name.date2="date2",
                        name.population.year1="pop1",
                        name.population.year2="pop2")
}
